requirements.txt
asyncio==3.4.3
RPi.GPIO==0.6.2
websockets==3.1
